requirements.txt
numpy>=1.17.4
torch
torchvision==0.2.1
PIL
pytorch-ignite
